added utils functions
