+++ src/containers/Flow/Flow.tsx
@@ -13,30 +13,9 @@ import { GET_ORGANIZATION } from 'graphql/queries/Organization';
 import { GET_FLOW } from 'graphql/queries/Flow';
 import { setErrorMessage } from 'common/notification';
 import Loading from 'components/UI/Layout/Loading/Loading';
-import { getUserSession } from 'services/AuthService';
-import { checkDynamicRole } from 'context/role';
 import styles from './Flow.module.css';
+import { getAddOrRemoveRoleIds } from 'common/utils';
 
-export const addOrRemoveRoles = (roles: any, payload: any) => {
-  const initialSelectedRoles = roles.map((role: any) => role.id);
-  const payloadRoleIds = payload.roles.map((role: any) => role.id);
-
-  let addRoleIds = payloadRoleIds.filter(
-    (selectedRoles: any) => !initialSelectedRoles.includes(selectedRoles)
-  );
-  const deleteRoleIds = initialSelectedRoles.filter(
-    (roleId: any) => !payloadRoleIds.includes(roleId)
-  );
-
-  if (checkDynamicRole()) {
-    const userRoles = getUserSession('roles').map((role: any) => role.id);
-    addRoleIds = [...addRoleIds, ...userRoles];
-  }
-
-  const { roles: userRoles, ...rest } = payload;
-
-  return { ...rest, addRoleIds, deleteRoleIds };
-};
 export interface FlowProps {
   match: any;
 }
@@ -180,11 +159,11 @@ export const Flow: React.SFC<FlowProps> = ({ match }) => {
       formattedKeywords = inputKeywords.split(',');
     }
 
-    const addRemoveRoles = addOrRemoveRoles(roles, payload);
+    const payloadWithRoleIds = getAddOrRemoveRoleIds(roles, payload);
 
     // return modified payload
     return {
-      ...addRemoveRoles,
+      ...payloadWithRoleIds,
       keywords: formattedKeywords,
     };
   };
